Feat: 구인글 유효성 검사

+++ src/main/java/com/example/bumerang/web/JobSearchController.java
@@ -1,5 +1,6 @@
 package com.example.bumerang.web;
 
+import com.example.bumerang.domain.comment.Comment;
 import com.example.bumerang.service.JobSearchService;
 import com.example.bumerang.web.dto.SessionUserDto;
 import com.example.bumerang.web.dto.request.jobSearch.UpdateDto;
@@ -23,52 +24,67 @@ public class JobSearchController {
     private final JobSearchService jobSearchService;
 
     // 구인정보 작성하기 화면
-    @GetMapping("/jobSearch/writeForm")
+    @GetMapping("/s/api/jobSearch/writeForm")
     public @ResponseBody CMRespDto<?> writeForm() {
         return new CMRespDto<>(1, "구인정보 작성하기 화면 불러오기 성공.", null);
     }
 
     // 구인정보 작성하기 기능
-    @PostMapping("/jobSearch/write")
+    @PostMapping("/s/api/jobSearch/write")
     public @ResponseBody CMRespDto<?> write(@RequestBody WriteDto writeDto) {
-        JobRespDto writeResult = jobSearchService.write(writeDto);
-        return new CMRespDto<>(1, "구인정보 작성하기 성공.", writeResult);
+        SessionUserDto principal = (SessionUserDto)session.getAttribute("principal");
+        Integer userId = writeDto.getUserId();
+        Integer userPId = principal.getUserId();
+        if(userId == userPId){
+            JobRespDto writeResult = jobSearchService.write(writeDto);
+            return new CMRespDto<>(1, "구인정보 작성하기 성공.", writeResult);
+        }
+        return new CMRespDto<>(-1, "올바르지 않은 요청입니다.", null);
     }
     
     // 구인정보 상세보기 화면
-    @GetMapping("/jobSearch/detailForm/{jobId}")
+    @GetMapping("/s/api/jobSearch/detailForm/{jobId}")
     public @ResponseBody CMRespDto<?> detailForm(@PathVariable Integer jobId, Model model) {
         SessionUserDto userPS = (SessionUserDto)session.getAttribute("principal");
-        if(userPS==null){
-            return new CMRespDto<>(-1, "로그인 해주세요.", null);
-        }
-        DetailFormDto jobDetail = jobSearchService.findByJob(userPS, jobId);
+        Integer userId = userPS.getUserId();
+        DetailFormDto jobDetail = jobSearchService.findByJob(userId, jobId);
         return new CMRespDto<>(1, "구인정보 상세보기 화면 불러오기 성공.", jobDetail);
     }
 
     // 구인정보 수정하기 화면
-    @GetMapping("/jobSearch/updateForm/{jobId}")
+    @GetMapping("/s/api/jobSearch/updateForm/{jobId}")
     public @ResponseBody CMRespDto<?> updateForm(@PathVariable Integer jobId) {
         SessionUserDto userPS = (SessionUserDto)session.getAttribute("principal");
-        if(userPS==null){
-            return new CMRespDto<>(-1, "로그인 해주세요.", null);
-        }
-        DetailFormDto jobDetail = jobSearchService.findByJob(userPS, jobId);
+        Integer userId = userPS.getUserId();
+        DetailFormDto jobDetail = jobSearchService.findByJob(userId, jobId);
         return new CMRespDto<>(1, "구인정보 수정하기 화면 불러오기 성공.", jobDetail);
     }
 
     // 구인정보 수정하기 기능
-    @PutMapping("/jobSearch/update")
+    @PutMapping("/s/api/jobSearch/update")
     public @ResponseBody CMRespDto<?> update(@RequestBody UpdateDto updateDto) {
-        JobRespDto updateResult = jobSearchService.update(updateDto);
-        return new CMRespDto<>(1, "구인정보 수정하기 성공.", updateResult);
+        SessionUserDto principal = (SessionUserDto)session.getAttribute("principal");
+        Integer userId = updateDto.getUserId();
+        Integer userPId = principal.getUserId();
+        if(userId == userPId){
+            JobRespDto updateResult = jobSearchService.update(updateDto);
+            return new CMRespDto<>(1, "구인정보 수정하기 성공.", updateResult);
+        }
+        return new CMRespDto<>(-1, "올바르지 않은 요청입니다.", null);
     }
 
     // 구인정보 삭제하기 기능
-    @DeleteMapping("/jobSearch/delete/{jobId}")
-    public @ResponseBody CMRespDto<?> delete(@PathVariable Integer jobId) {
-        JobRespDto deleteResult = jobSearchService.delete(jobId);
-        return new CMRespDto<>(1, "구인정보 삭제하기 성공.", deleteResult);
+    @DeleteMapping("/s/api/jobSearch/delete")
+    public @ResponseBody CMRespDto<?> delete(@RequestBody DetailFormDto detailFormDto) {
+        SessionUserDto principal = (SessionUserDto)session.getAttribute("principal");
+        Integer userId = detailFormDto.getUserId();
+        Integer jobId = detailFormDto.getJobId();
+        Integer userPId = principal.getUserId();
+        if(userId == userPId){
+            JobRespDto deleteResult = jobSearchService.delete(jobId);
+            return new CMRespDto<>(1, "구인정보 삭제하기 성공.", deleteResult);
+        }
+        return new CMRespDto<>(-1, "올바르지 않은 요청입니다.", null);
     }
 
 }
